qwen-target/pyproject.toml
[project]
name = "qwen-target"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "transformers>=4.57.3",
    "playwright>=1.49.1",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "ipdb>=0.13.13",
    "pydantic>=2.12.5",
]
